feat : 이벤트 삭제
